requirements.txt
aws-cdk-lib
constructs>=10.0.0
boto3>=1.26.0
requests>=2.28.0
beautifulsoup4>=4.11.0
constructs>=10.0.0,<11.0.0
